Fix subj and prof component names
