+++ client/src/ui/SubjectResult.tsx
@@ -1,17 +1,17 @@
 import React, { Component} from 'react';
 import "./css/Course.css";
 import { Redirect } from 'react-router';
-import { Professor } from 'common';
+import { Subject } from 'common';
 
 /*
-  Professor Component.
+  Subject Component.
 
-  Simple styling component that represents a single professor (an a element).
-  Used to list professor in the results of a search in SearchBar.
+  Simple styling component that represents a single subject (an a element).
+  Used to list subject in the results of a search in SearchBar.
 */
 
 type Props = {
-  professor: Professor;
+  info: Subject;
   query?: string; //optional
   active: boolean;
   enter: 1 | 0;
@@ -19,15 +19,14 @@ type Props = {
   key: string;
 };
 
-export default class Course extends Component<Props> {
+export default class SubjectResult extends Component<Props> {
   render() {
     // generate full human-readable name of class
-    const professorObject = this.props.professor;
-    let text = professorObject.fullName;
-    let professorURL = professorObject.fullName.split(" ").join("+")
+    const subjectInfo = this.props.info;
+    let text = subjectInfo.subFull;
     //if the element is highlighted and the enter key was pressed, create a Redirect component to go to the class
     if(this.props.active && this.props.enter === 1){
-       return <Redirect push to={`/results/professor/${professorURL}`}></Redirect>
+       return <Redirect push to={`/results/major/${subjectInfo.subShort.toUpperCase()}`}></Redirect>
       }
 
     //return classname as a list element
@@ -36,9 +35,9 @@ export default class Course extends Component<Props> {
     //if the mouse is in the list element, highlighting by arrow key stops and follow the mouse hovers
     //if the mouse leaves the list element, highlighting by arrow key continues but from the first element
       <a className={this.props.active && this.props.mouse !== 1 ? 'active-class resultbutton' : 'resultbutton'}
-          href={`/results/professor/${professorURL}`}>
-        <p className="result-label-professor">
-          Professor
+          href={`/results/major/${subjectInfo.subShort.toUpperCase()}`}>
+        <p className="result-label-subject">
+          Major
         </p>
         <p className="result-text">{text}</p>
       </a>
